Extract file receiver from `BackupManager.async_receive_backup` to util (#132271)
* Extract file receiver from BackupManager.async_receive_backup to util

* Apply suggestions from code review


---------
