+++ homeassistant/components/backup/util.py
@@ -2,10 +2,15 @@
 
 from __future__ import annotations
 
+import asyncio
 from pathlib import Path
+from queue import SimpleQueue
 import tarfile
 from typing import cast
 
+import aiohttp
+
+from homeassistant.core import HomeAssistant
 from homeassistant.util.json import JsonObjectType, json_loads_object
 
 from .const import BUF_SIZE
@@ -57,3 +62,45 @@ def read_backup(backup_path: Path) -> AgentBackup:
             protected=cast(bool, data.get("protected", False)),
             size=backup_path.stat().st_size,
         )
+
+
+async def receive_file(
+    hass: HomeAssistant, contents: aiohttp.BodyPartReader, path: Path
+) -> None:
+    """Receive a file from a stream and write it to a file."""
+    queue: SimpleQueue[tuple[bytes, asyncio.Future[None] | None] | None] = SimpleQueue()
+
+    def _sync_queue_consumer() -> None:
+        with path.open("wb") as file_handle:
+            while True:
+                if (_chunk_future := queue.get()) is None:
+                    break
+                _chunk, _future = _chunk_future
+                if _future is not None:
+                    hass.loop.call_soon_threadsafe(_future.set_result, None)
+                file_handle.write(_chunk)
+
+    fut: asyncio.Future[None] | None = None
+    try:
+        fut = hass.async_add_executor_job(_sync_queue_consumer)
+        megabytes_sending = 0
+        while chunk := await contents.read_chunk(BUF_SIZE):
+            megabytes_sending += 1
+            if megabytes_sending % 5 != 0:
+                queue.put_nowait((chunk, None))
+                continue
+
+            chunk_future = hass.loop.create_future()
+            queue.put_nowait((chunk, chunk_future))
+            await asyncio.wait(
+                (fut, chunk_future),
+                return_when=asyncio.FIRST_COMPLETED,
+            )
+            if fut.done():
+                # The executor job failed
+                break
+
+        queue.put_nowait(None)  # terminate queue consumer
+    finally:
+        if fut is not None:
+            await fut
